style: Rename api.ts to api-handler.ts types
